# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ToolKit API",
        "version": "1.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "operationId": "eb3623f9bcd09a1dacbefaac07aed1bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "number": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "f77777235d676b651c9465807d8f9485",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/auth/profile": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Profile",
                "operationId": "8e43034ac8f3aaaf5c015750b446596b",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "c4b3c6f1776909e546a421283c044ac3",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh",
                "operationId": "a11d4212c7e49cbf5852d9bb2e3d5054",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/declaration": {
            "get": {
                "tags": [
                    "Declaration"
                ],
                "summary": "All declarations",
                "operationId": "1e69a9f2f13ace980593a0f7228b81a9",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Доклад №1"
                                            },
                                            "date": {
                                                "type": "string",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "email_verified_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Declaration"
                ],
                "summary": "Declaration",
                "operationId": "a6161c2e55e39c05cdcac3883e218fd6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/declaration/{id}": {
            "delete": {
                "tags": [
                    "Declaration"
                ],
                "summary": "Delete",
                "operationId": "4b8fe3de69e1acd1a6d9ff0b2482cf39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Declaration"
                ],
                "summary": "Patch",
                "operationId": "342f9dacb2ab7fa21a78a0ed23751079",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "All users",
                "operationId": "74a51985aa8ca65de0bab1b1065a2f3f",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Самса"
                                            },
                                            "email_verified_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "89999999999"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "838 Destinee Lodge\\nKuhicfort, OK 90655"
                                            },
                                            "is_admin": {
                                                "type": "boolean",
                                                "example": "false"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete",
                "operationId": "6c44fd76baabb897f45289fd52179b57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Patch",
                "operationId": "eb2cf2407bd42a8024096b55ce14538e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "number": {
                                        "type": "string",
                                        "format": "number"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "address"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Declaration",
            "description": "Declaration"
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}